update to entity based prototypes

+++ Content.Server/_CD/Silicons/Borgs/BorgSwitchableSubstypeSystem.cs
@@ -1,7 +1,5 @@
-using Content.Shared._CD.Silicons;
 using Content.Shared._CD.Silicons.Borgs;
 using Content.Shared.Inventory;
-using Robust.Shared.Prototypes;
 
 namespace Content.Server._CD.Silicons.Borgs;
 
@@ -17,12 +15,15 @@ public sealed class BorgSwitchableSubstypeSystem : SharedBorgSwitchableSubtypeSy
         if (ent.Comp.BorgSubtype == null)
             return;
 
-        var borgSubtype = Prototypes.Index(ent.Comp.BorgSubtype.Value);
+        if (!Prototypes.Index(ent.Comp.BorgSubtype.Value)
+            .TryGetComponent<BorgSubtypeDefinitionComponent>(out var borgSubtype, ComponentFactory))
+            return;
 
         // Configure special components
-        if (Prototypes.TryIndex(ent.Comp.BorgSubtype, out var previousPrototype))
+        if (Prototypes.TryIndex(ent.Comp.BorgSubtype, out var previousPrototype) &&
+            previousPrototype.TryGetComponent<BorgSubtypeDefinitionComponent>(out var previousSubtype, ComponentFactory))
         {
-            if (previousPrototype.AddComponents is { } removeComponents)
+            if (previousSubtype.AddComponents is { } removeComponents)
                 EntityManager.RemoveComponents(ent, removeComponents);
         }
 
@@ -37,7 +38,6 @@ public sealed class BorgSwitchableSubstypeSystem : SharedBorgSwitchableSubtypeSy
             _inventorySystem.SetTemplateId((ent.Owner, inventory), borgSubtype.InventoryTemplateId);
         }
 
-        Dirty(ent);
         base.SelectBorgSubtype(ent);
     }
 }
